fix linux distribution build (Ubuntu) by allow -DDISTRIBUTION_BUILD=1
This adds the cmake option "-DDISTRIBUTION_BUILD=1" to inform add-on build
about Linux Distribution build.
This becomes needed as on Ubuntu used libusb becomes linked static, but not created
as "position independent code". With the option becomes the linkage about prevented,
as the lib already as dependency in librtlsdr becomes it to add-on as shared on this
way added.