configure settings.xml during build to define USB not supported
This change it to settings.xml.in and set on add-on compile the settings usable or not.
Is done by change the `<visible>` value and set of default connection type.
The settings are still inside to prevent problems if something tries somewhere to read it.

Further can for build the cmake configuration value "-DUSB_DEVICE_SUPPORT=0" used and if wanted to disable USB support.